pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "httpx==0.27.0",
    "nest-asyncio>=1.6.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv==1.1.1",
    "python-telegram-bot==21.9",
]
